requirements.txt
torch
torchvision
numpy
pandas
opencv-python
tqdm
Pillow
